chore: update Docker ignore and Git ignore files, modify Docker Compose configuration
This commit updates the `.dockerignore` and `.gitignore` files to include additional patterns for ignoring environment files and IDE configurations. It also modifies the `compose.yaml` file to change the default value for the `DB_OWNER` environment variable and updates the port mapping for the TN RPC service, enhancing the clarity and maintainability of the Docker setup. Additionally, the `Taskfile.yml` is updated to include a new `PROVIDER` environment variable for migration tasks.