feat: add mainnet.truf.network hosted zone configuration and simplify SAN addition logic
This commit introduces a new hosted zone configuration for the domain `mainnet.truf.network` in the CDK context file. Additionally, it simplifies the logic for adding Subject Alternative Names (SANs) in the `AlternativeDomainManager` by removing the conditional check for the existence of target components, ensuring that SANs are added unconditionally when required. These changes enhance the domain management capabilities and streamline the certificate building process.