chore: add observer service to CDK deployment (#957)
* feat: add block synchronization timeouts and observer systemd service

This commit introduces several timeout configurations for block synchronization, including timeouts for fetching, sending, and processing blocks. Additionally, a systemd service script for the observer component is created to manage its lifecycle effectively. These enhancements improve the observability and performance of the node's operations.

* fix: update SSM prefix in AttachObservability function

This commit modifies the SSM prefix in the AttachObservability function to a more general path, simplifying the configuration for observability attachment. The change enhances clarity and maintainability of the code.

* feat: add AwsRegion field to observer scripts and templates

This commit introduces the AwsRegion field to various observer-related structures and templates, enhancing the configurability of the observer component. The changes include updates to the AttachObservability function, observer script input types, and the observer start template, ensuring that the AWS region can be specified and utilized effectively. Additionally, tests have been updated to reflect these changes, improving the overall robustness of the implementation.

* fix: update observer asset directory path

This commit modifies the path for the observer assets directory in the GetObserverAsset function, changing it from "deployments/infra/lib/observer" to "deployments/observer". This change improves the organization of the asset files and enhances clarity in the codebase.

* fix: update service name format in AttachObservability function

This commit modifies the ServiceName format in the AttachObservability function to include the instance index and a version suffix, changing it from "tn-node" to "tn-node-{index}-v2". This change enhances clarity and consistency in service naming conventions.