chore(deps): update kwil-db and go-ethereum dependencies in go.mod and go.sum
This commit updates the versions of `kwil-db` and `kwil-db/core` in the `go.mod` and `go.sum` files to the latest revisions. Additionally, it ensures that the `go-ethereum` dependency is correctly listed without the indirect comment. These changes aim to keep the project dependencies up to date and maintain compatibility with the latest features and fixes.