chore: add GitHub Actions workflow for publishing Node image (#1166)
* feat: add GitHub Actions workflow for publishing Node image

This commit introduces a new workflow in `.github/workflows/publish-node-image.yaml` to automate the building and pushing of a Docker image to the GitHub Container Registry upon release events. Additionally, it updates the `Dockerfile` to use environment variables for configuration and refines the entrypoint script to improve clarity and maintainability.

* feat: enhance GitHub Actions workflow for Node image publishing

This commit adds an input option to the workflow for tagging builds as 'latest' during manual dispatch. It also introduces an environment variable to control the tagging logic based on the event type, improving the flexibility of the image publishing process.

* feat: update GitHub Actions workflow to support multi-architecture builds

This commit modifies the Node image publishing workflow to enable building and pushing Docker images for both amd64 and arm64 architectures. This enhancement improves compatibility and accessibility of the published images across different platforms.